logging in redirects to account screen, which only has log out functionality for now, which takes you back to map screen. Moved style files into screens folder, deleted sub-screen folders.
