+++ App.js
@@ -5,9 +5,9 @@ import 'react-native-gesture-handler';
 import { firebase } from './src/firebase/config'
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
-import LoginScreen from './src/screens/LoginScreen/LoginScreen';
-import HomeScreen from './src/screens/HomeScreen/HomeScreen';
-import RegistrationScreen from './src/screens/RegistrationScreen/RegistrationScreen';
+import LoginScreen from './src/screens/LoginScreen';
+import AccountScreen from './src/screens/AccountScreen';
+import RegistrationScreen from './src/screens/RegistrationScreen';
 import MapScreen from './src/screens/MapScreen';
 import {decode, encode} from 'base-64'
 if (!global.btoa) {  global.btoa = encode }
@@ -58,7 +58,7 @@ export default function App() {
     <NavigationContainer>
       <Stack.Navigator>
         { user ? (
-          <Stack.Screen name="Map" component={MapScreen}>
+          <Stack.Screen name="Account" component={AccountScreen}>
             {props => <MapScreen {...props} extraData={user} />}
           </Stack.Screen>
         ) : (
@@ -66,7 +66,7 @@ export default function App() {
             <Stack.Screen name="Map" component={MapScreen}/> 
             <Stack.Screen name="Login" component={LoginScreen} />
             <Stack.Screen name="Registration" component={RegistrationScreen} /> 
-            <Stack.Screen name="Home" component={HomeScreen}/> 
+            <Stack.Screen name="Account" component={AccountScreen}/> 
           </>
         )}
       </Stack.Navigator>
