+++ index.js
@@ -1,8 +1,8 @@
 import { registerRootComponent } from 'expo';
-export { default as LoginScreen } from './src/screens/LoginScreen/LoginScreen'
-export { default as HomeScreen } from './src/screens/HomeScreen/HomeScreen'
+export { default as LoginScreen } from './src/screens/LoginScreen'
+export { default as AccounteScreen } from './src/screens/AccountScreen'
 export { default as MapScreen } from './src/screens/MapScreen'
-export { default as RegistrationScreen } from './src/screens/RegistrationScreen/RegistrationScreen'
+export { default as RegistrationScreen } from './src/screens/RegistrationScreen'
 import App from './App';
 
 // registerRootComponent calls AppRegistry.registerComponent('main', () => App);
