+++ src/components/StarRating.js
