--- src/screens/HomeScreen/HomeScreen.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import { Text, View } from 'react-native'
-
-export default function HomeScreen(navigation) {
-    return (
-        <View>
-            <Text>Home Screen after login (Need to insert main map screen here)</Text>
-        </View>
-    )
-}
