--- src/screens/index.js
@@ -1,1 +1,0 @@
-//used in root index.js file instead
