+++ src/screens/registrationStyles.js
